Reducing memory foot print and in between storing on client and server side

+++ src/modules/NuFFT.hpp
@@ -75,7 +75,7 @@ namespace RRStrategy {
 		int*      m_n;                          /**< @brief Oversampling */
 		int       m_M;                          /**< @brief Number of k-space knots */
 		int       m_maxit;                      /**< @brief Number of Recon iterations (NFFT 3) */
-		int       m_epsilon;                    /**< @brief Convergence criterium */
+		double    m_epsilon;                    /**< @brief Convergence criterium */
 		int       m_shots;
 
 		double*   m_ftk;
